requirements.txt
asgiref==3.8.1
Django==5.1.4
django-axes==7.0.1
django-cors-headers==3.13.0
django-filter==24.3
django-ipware==4.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
